+++ product/src/database/inMemory/inMemory.module.ts
@@ -1,6 +1,6 @@
 import { Module } from "@nestjs/common";
-import { AppMemoryModule } from "../appMemory/appMemory.module";
-import { RedisModule } from "../redis/redis.module";
+import { AppMemoryModule } from "./appMemory/appMemory.module";
+import { RedisModule } from "./redis/redis.module";
 
 @Module({
     imports: [
